+++ example/example.dart
@@ -5,12 +5,18 @@ void main() {
   generateFile(File('generated.dart'), library);
 }
 
-final library = Library(
-  dynamicLibraryPath: 'path/to/library',
+final library = const Library.platformAware(
+  // Configure, how the dynamic library should be loaded depending on the platform
+  dynamicLibraryConfig: DynamicLibraryConfig(
+      windows: DynamicLibraryPlatformConfig.open('path/to/library.dll'),
+      android: DynamicLibraryPlatformConfig.open('path/to/library.so'),
+      iOS: DynamicLibraryPlatformConfig.process()),
+
   elements: [
     // A function
     Func(
       name: 'Example',
+      documentation: 'Takes parameters and does stuff.',
       parameterTypes: ['int32', 'float64', '*void'],
       returnType: 'void',
     ),