+++ lib/src/c/global.dart
@@ -18,21 +18,30 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 // OR OTHER DEALINGS IN THE SOFTWARE.
 
-import 'package:ffi_tool/c.dart';
 import 'package:meta/meta.dart';
+
+import 'library.dart';
+import 'dart_source_writer.dart';
 
 /// A definition for a C global.
 class Global extends Element {
-  final String name;
   final String type;
-  const Global({@required this.name, @required this.type});
+  const Global(
+      {@required String name, @required this.type, String documentation})
+      : super(name: name, documentation: documentation);
 
   @override
   void generateSource(DartSourceWriter w, Library library) {
     final dartType = w.getDartType(type);
     final cType = w.getCType(type);
     w.write('\n');
-    w.write('/// C global `$name`.\n');
+        if (documentation == null) {
+      w.write('/// C global `$name`.\n');
+    } else {
+      w.write('/// ');
+      w.writeAll(documentation.split('\n'), '\n/// ');
+      w.write('\n');
+    }
     w.write(
         'final $dartType $name = ${library.dynamicLibraryIdentifier}.lookup<$cType>(\n');
     w.write('  \'$name\',\n');